Change `useRouteTypedFields` to a synchronous util
