+++ packages/react-native/src/Authenticator/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './useFieldValues';
-export * from './useRouteTypedFields';
 export * from './types';
