+++ packages/react-native/src/Authenticator/hooks/useFieldValues/index.ts
@@ -1,1 +1,2 @@
 export { default as useFieldValues } from './useFieldValues';
+export { getRouteTypedFields } from './utils';
