+++ packages/react-native/src/Authenticator/hooks/useFieldValues/utils.ts
@@ -1,7 +1,18 @@
 import { Logger } from 'aws-amplify';
-import { AuthenticatorRouteComponentName } from '@aws-amplify/ui-react-core';
+import {
+  AuthenticatorLegacyField,
+  AuthenticatorRouteComponentName,
+  isAuthenticatorComponentRouteKey,
+  UseAuthenticator,
+} from '@aws-amplify/ui-react-core';
 
-import { RadioFieldOptions, TypedField } from '../types';
+import {
+  AuthenticatorFieldTypeKey,
+  MachineFieldTypeKey,
+  RadioFieldOptions,
+  TypedField,
+} from '../types';
+import { KEY_ALLOW_LIST } from './constants';
 
 const logger = new Logger('Authenticator');
 
@@ -78,3 +89,85 @@ export const getSanitizedTextFields = (
     return true;
   });
 };
+
+// typed fields utils
+const isKeyAllowed = (key: string) =>
+  KEY_ALLOW_LIST.some((allowedKey) => allowedKey === key);
+
+const isValidMachineFieldType = (
+  type: string | undefined
+): type is MachineFieldTypeKey => type === 'password' || type === 'tel';
+
+const getFieldType = (type: string | undefined): AuthenticatorFieldTypeKey => {
+  if (isValidMachineFieldType(type)) {
+    return type === 'tel' ? 'phone' : type;
+  }
+  return 'default';
+};
+
+/**
+ * Translate machine fields to typed fields
+ *
+ * @param {AuthenticatorLegacyField} field machine field option object
+ * @returns {TypedField} UI field props object
+ */
+export const getTypedField = ({
+  type: machineFieldType,
+  name,
+  ...field
+}: AuthenticatorLegacyField): TypedField => {
+  const type = getFieldType(machineFieldType);
+
+  return Object.entries(field).reduce(
+    (acc, [key, value]) => {
+      // early return if key is not allowed
+      if (!isKeyAllowed(key)) {
+        return acc;
+      }
+
+      // map to `required` prop
+      if (key === 'isRequired' || key === 'required') {
+        // `TypedField` props expects `required` key
+        return { ...acc, required: value as boolean };
+      }
+
+      const isPhoneFieldType = type === 'phone';
+
+      if (isPhoneFieldType && key === 'dialCode') {
+        return { ...acc, defaultDialCode: value as string };
+      }
+
+      if (isPhoneFieldType && key === 'dialCodeList') {
+        return { ...acc, dialCodes: value as string[] };
+      }
+
+      return { ...acc, [key]: value };
+    },
+    // initialize `acc` with field `name` and `type`
+    { name, type } as TypedField
+  );
+};
+
+/**
+ * @param {AuthenticatorLegacyField[]} fields machine field options array
+ * @returns {TypedField[]} UI field props array
+ */
+export const getTypedFields = (
+  fields: AuthenticatorLegacyField[]
+): TypedField[] => fields?.map(getTypedField);
+
+export function getRouteTypedFields({
+  fields,
+  route,
+}: Pick<UseAuthenticator, 'fields' | 'route'>): TypedField[] {
+  const isComponentRoute = isAuthenticatorComponentRouteKey(route);
+  if (isComponentRoute) {
+    // `VerifyUser` does not require additional updates to the shape of `fields`
+    const isVerifyUserRoute = route === 'verifyUser';
+    const radioFields = fields as TypedField[];
+
+    return isVerifyUserRoute ? radioFields : getTypedFields(fields);
+  } else {
+    return [];
+  }
+}
