--- packages/react-native/src/Authenticator/hooks/useRouteTypedFields/index.ts
@@ -1,1 +1,0 @@
-export { default as useRouteTypedFields } from './useRouteTypedFields';
